Add CalculatorDisplay component and style within Display.css.
