+++ Project-II-Calculator/calculator/src/components/DisplayComponents/Display.css
@@ -1,4 +1,13 @@
 .calculator-display {
-    margin: 0 auto;
-    max-width: 400px;
+    display: flex;
+    justify-content: flex-end;
+    padding-right: .5rem;
+    align-items: center;
+    max-width: 405px;
+    border-bottom: 2px solid #979797;
+    height: 110px;
+    font-size: 4rem;
+    font-weight: lighter;
+    background: #222222;
+    color: white;
 }
